public-blog/infra/requirements.txt
aws-cdk-lib==2.15.0
constructs>=10.0.0,<11.0.0
requests==2.27.1
pyyaml==6.0
